refactor(outbound): extract runtime and session tests
